feat: Add search functionality for meals by name
The code changes in this commit add a search input field to the "Ingerdient-meal.vue" component, allowing users to search for meals by name. The search functionality filters the list of ingredients based on the keyword entered in the search input.

This commit also includes updates to the "Meals-Letter.vue" and "Meals-by-name.vue" components to display the filtered meals based on the search keyword.

Recent user commits:
- added meal by name, ingredient
- updated nav-bar routes, meal-details

Recent repository commits:
- getting meals by letter
- added mealItem
- refactor: Update mutation.js to use named export for setSearchedMeals
- chore: Add Tailwind CSS, Vue Router, and Vuex to project
- init